+++ crates/rethnet_eth/src/remote/eth.rs
@@ -6,6 +6,8 @@
 // - https://github.com/gakonst/ethers-rs/blob/7e6c3ba98363bdf6131e8284f186cc2c70ff48c3/LICENSE-MIT
 // For the original context, see https://github.com/gakonst/ethers-rs/tree/7e6c3ba98363bdf6131e8284f186cc2c70ff48c3
 
+pub mod eip712;
+
 use std::fmt::Debug;
 
 use crate::{Address, Bloom, Bytes, B256, U256};
@@ -174,44 +176,4 @@ where
     use serde::Deserialize;
     let opt = Option::deserialize(deserializer)?;
     Ok(opt.unwrap_or_default())
-}
-
-pub mod eip712 {
-    // adapted from https://github.com/openethereum/parity-ethereum/blob/v2.7.2-stable/util/EIP-712/src/eip712.rs
-
-    use super::*;
-    use hashbrown::HashMap;
-
-    #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq)]
-    pub struct FieldType {
-        pub name: String,
-        #[serde(rename = "type")]
-        pub type_: String,
-    }
-    pub type MessageTypes = HashMap<String, Vec<FieldType>>;
-
-    #[derive(serde::Deserialize, serde::Serialize, Debug, Clone, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub struct Domain {
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub name: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub version: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub chain_id: Option<U256>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub verifying_contract: Option<Address>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub salt: Option<B256>,
-    }
-
-    #[derive(serde::Deserialize, serde::Serialize, Debug, Clone, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    #[serde(deny_unknown_fields)]
-    pub struct Message {
-        pub types: MessageTypes,
-        pub primary_type: String,
-        pub message: serde_json::Value,
-        pub domain: Domain,
-    }
 }
